refactor: 3단계 - 경로 조회 기능 (PR 피드백 2 적용)
- 코드 정리
- pathfinder 로직 개선

+++ src/main/java/nextstep/subway/applicaion/dto/path/PathResponse.java
@@ -8,9 +8,9 @@ import java.util.stream.Collectors;
 
 public class PathResponse {
     private List<StationResponse> stations;
-    private int distance = 0;
+    private double distance = 0;
 
-    public PathResponse(List<StationResponse> stations, int distance) {
+    public PathResponse(List<StationResponse> stations, double distance) {
         this.stations = stations;
         this.distance = distance;
     }
@@ -27,7 +27,7 @@ public class PathResponse {
         return stations;
     }
 
-    public int getDistance() {
+    public double getDistance() {
         return distance;
     }
 }
